requirements.txt
black==25.1.0
click==8.2.1
isort==6.0.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pyzmq==27.0.0
